model agnostic part of the workflow using hydromt datacatalogue
Hydromt allows reading all raster data into xarrays, which makes uniform and in memory processing possible. It eleminates reliance on the qgis plugin and separates the model-specific part from the model-agnostic part.

The example to run is bow_at_banff_model_agnostic. It still retains the file structure from CWARHM (as given in the control file), but saving intermediate steps is optional.

Two functions are added to HRU.py
1) add_basin_attributes_to_zonal_stats
2) count_per_class